# api-specification.yaml
openapi: 3.0.3
info:
  title: DCR Open Platform APIs
  description: APIs of the DCR Open Platform as developed at DTU.
  contact:
    email: andbur@dtu.dk
  version: 0.0.1

servers:
  - url: http://localhost:8080/api/v1/
    description: Development
  - url: https://dcr-engine-eu1.herokuapp.com/api/v1/
    description: Deployed

paths:
  /system/ping:
    get:
      tags:
        - system
      summary: Ping the system
      description: Checks if the system is online
      responses:
        200:
          description: The system is up and running
  /dcr/dcr2graphviz:
    post:
      tags:
        - visualization
      summary: Converts a process into Graphviz
      description: Generates a Graphviz representation of the process.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Process'
      responses:
        200:
          description: The process has been correctly rendered
          content:
            plain/text:
              schema:
                type: string
  
  /dcr/simulation/initialize:
    post:
      tags:
        - simulation
      summary: Initializes the simulation of the process
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Process'
      responses:
        200:
          description: The simulation has been initialized correctly and the simulation id is returned
          content:
            plain/text:
              schema:
                type: string
                format: uuid
  /dcr/simulation/{simulationId}/status:
    get:
      tags:
        - simulation
      summary: Returns the status of the simulation
      parameters:
        - in: path
          name: simulationId
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the simulation
      responses:
        404:
          description: The required simulation is not found
        200:
          description: The status of the required simulation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationStatus'
  /dcr/simulation/{simulationId}/reset:
    get:
      tags:
        - simulation
      summary: Removes the simulation from the system
      parameters:
        - in: path
          name: simulationId
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the simulation
      responses:
        200:
          description: The simulation is removed from the sistem
  /dcr/simulation/{simulationId}/execute/{activityId}:
    get:
      tags:
        - simulation
      summary: Execute the activity in the given simulation
      parameters:
        - in: path
          name: simulationId
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the simulation
        - in: path
          name: activityId
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the activity to execute
      responses:
        404:
          description: The required simulation is not found
        200:
          description: The status of the simulation afterthe activity has been executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationStatus'
  /dcr/edit/add/{sourceActivity}/{relation}/{targetActivity}:
    post:
      tags:
        - editor
      summary: Removes the simulation from the system
      parameters:
        - in: path
          name: sourceActivity
          schema:
            type: string
          required: true
          description: The name of the source activity
        - in: path
          name: relation
          schema:
            type: string
            enum: [CONDITION, RESPONSE, INCLUDE, EXCLUDE, MILESTONE]
          required: true
          description: The name of the relation
        - in: path
          name: targetActivity
          schema:
            type: string
          required: true
          description: The name of the target activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Process'
      responses:
        200:
          description: The simulation is removed from the sistem

components:
  schemas:
    Process:
      type: object
      properties:
        activities:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        relations:
          type: array
          items:
            $ref: '#/components/schemas/Relation'
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Relation:
      type: object
      properties:
        source:
          type: string
          format: uuid
        target:
          type: string
          format: uuid
        relation:
          type: string
          enum: [CONDITION, RESPONSE, INCLUDE, EXCLUDE, MILESTONE]
    
    SimulationStatus:
      type: object
      properties:
        enabledActivities:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        isAccepting:
          type: boolean